feat: add usersList
