+++ src/app/components/userPage.jsx
@@ -1,52 +1,18 @@
 import React from 'react';
-import PropTypes from 'prop-types';
-import BookMark from './bookmark';
-import Qualitie from './qualitie';
+import { Link } from 'react-router-dom';
+import { useHistory } from 'react-router';
+// import API from '../api';
 
-const UserPage = ({
-  name,
-  qualities,
-  profession,
-  completedMeetings,
-  rate,
-  _id,
-  onDelete,
-  onToggleBookMark,
-  status
-}) => {
+const userPage = ({ user }) => {
+  const history = useHistory();
+  console.log(history);
+  // const userId = API.users.getById(id);
+  // console.log(userId);
   return (
-    <tr>
-      <th>{name}</th>
-      <td>
-        {qualities.map((item) => (
-          <Qualitie key={item._id} {...item} />
-        ))}
-      </td>
-      <td>{profession.name}</td>
-      <td>{completedMeetings}</td>
-      <td>{rate}/5</td>
-      <td>
-        <BookMark id={_id} onToggleBookMark={onToggleBookMark} status={status} />
-      </td>
-      <td>
-        <button className="btn btn-danger" onClick={() => onDelete(_id)}>
-          Delete
-        </button>
-      </td>
-    </tr>
+    <>
+      <Link to={`/${user._id}`}>{user.name}</Link>
+    </>
   );
 };
 
-UserPage.propTypes = {
-  name: PropTypes.string,
-  qualities: PropTypes.array,
-  profession: PropTypes.object,
-  completedMeetings: PropTypes.number,
-  rate: PropTypes.number,
-  _id: PropTypes.string,
-  onDelete: PropTypes.func,
-  onToggleBookMark: PropTypes.func,
-  status: PropTypes.bool
-};
-
-export default UserPage;
+export default userPage;
